Move test helper used in one place to test class
